Move event metadata functions to ssp_read_event_metadata.py
